+++ tests/Integration/Internal/Transition/Smarty/SmartyBuilderTest.php
@@ -42,7 +42,6 @@ class SmartyBuilderTest extends \PHPUnit\Framework\TestCase
      */
     public function testSmartySettingsAreSetCorrect($securityMode, $smartySettings)
     {
-        $this->markTestSkipped('TODO: Refactor');
         /** @var SmartyConfigurationFactory $configurationFactory */
         $configurationFactory = $this->setupAndConfigureContainer($securityMode)
             ->get(SmartyConfigurationFactoryInterface::class);